requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
birdnetlib==0.3.1
black==23.1.0
boto3==1.26.78
botocore==1.29.78
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
darker==1.7.0
decorator==5.1.1
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
lazy_loader==0.1
libclang==15.0.6.1
librosa==0.10.0
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
msgpack==1.0.4
mypy-extensions==1.0.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
watchdog==2.1.9
Werkzeug==2.2.3
wrapt==1.14.1
